about files
there is text file you should put your path of this file to run the program and avoid any error so, you should to update the path. the function of the guessing text file is to store your name and your attempts and the standard deviations of the attempts of the user.